clean version

+++ app/src/main/java/com/example/themoviedb/database/MovieDatabaseDao.kt
@@ -5,7 +5,6 @@ import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.Query
 import com.example.themoviedb.model.Movie
-import com.example.themoviedb.model.MovieDetails
 
 @Dao
 interface MovieDatabaseDao {
@@ -23,5 +22,5 @@ interface MovieDatabaseDao {
 
 
     @Query("SELECT EXISTS(SELECT * from movies WHERE id = :id)")
-    suspend fun isFavorite(id: Int):Boolean
+    suspend fun isFavorite(id:Long):Boolean
 }
